Move the various Ctx's to their own file
Housekeeping motivated by the desire to mock out these types, and
wanting to have that all be in one place.
